(__pthread_manager): Do not terminate manager after unblocking main thread; wait for main thread's REQ_PROCESS_EXIT request instead. Also, added REQ_KICK case to handle new request; this just does nothing.
(pthread_exited): Do not terminate manager after unblocking main
thread.

(__pthread_manager_sighandler): If the main thread is waiting for all
other threads to die, send a REQ_KICK into the thread manager request
pipe to get it to clean out the threads and unblock the main thread as
soon as possible. This fixes the 2000 millisecond hang on shutdown
bug.