(Formatting Calendar Time): ISO C99 also specifies the E and O modifiers.
%P is a GNU extension, and is not in ISO C99.

Mention that %r is equivalent to %I:%M:%S %p in the POSIX locale.

%T is also in ISO C99.

The RFC 822 example is not valid in arbitrary locales.

Reword the POSIX.2 wording slightly, to make it a bit clearer
that POSIX.2 formats are also supported by later POSIX
versions.

If a format was introduced in ISO C99 it is also required by POSIX.1-2001.